refactor(environment variables): centralize variable value equivalence helper
